Solidity Feedback part 1 (#34)
* Solidity Feedback part 1

* self review

* get tests to run on PRs

* revert contracts yaml deletion

+++ packages/contracts/contracts/TestToken.sol
@@ -4,17 +4,15 @@ pragma solidity ^0.8.4;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract TestToken is ERC20 {
-    constructor() ERC20('Cabin Token', "CABIN") {
-        _mint(address(this), 10000 * 10**uint(decimals()));
+    constructor() ERC20("Cabin Token", "CABIN") {
+        _mint(address(this), 10000 * 10**uint256(decimals()));
     }
 
-    event Fauceted(
-        uint256 quantity
-    );
-    function faucet(uint256 quantity) public
-    {
+    event Fauceted(uint256 quantity);
+
+    function faucet(uint256 quantity) public {
         _approve(address(this), msg.sender, quantity);
         transferFrom(address(this), msg.sender, quantity);
-        emit Fauceted(quantity); 
+        emit Fauceted(quantity);
     }
 }
